delete image with delete product

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const mongoose = require('mongoose');
 
 const { validationResult } = require('express-validator/check');
+const fileHelper = require('../utils/file');
 
 const Product = require('../models/product');
 
@@ -159,6 +160,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then((result) => {
@@ -194,7 +196,13 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) return next(new Error('Product not found'));
+      fileHelper.deleteFile(product.imageUrl);
+      return 
+      Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
